win32: fix most of the compile warnings

+++ src/thread/win32/SDL_syssem.c
@@ -38,7 +38,7 @@ struct SDL_semaphore {
 #else
 	HANDLE id;
 #endif
-	Uint32 volatile count;
+	LONG volatile count;
 };
